Make tests more robust
The tests should be a little more conservative in the kind of types they
define as doing that affects the code generation of the plugin itself.
In this case, the type was introduced just for the show instance. So we
import the show instance for error_or in a way that doesn't pollute the
environment.