fix behavior of %of_yojson, %to_yojson on types with type variables
Testing for the `safe_wrap` issue revealed another issue with the
%of_yojson, %to_yojson cases: incorrect handling of type with
parameters `'a option`, which would generate calls to unbound `poly_a`
parameters. For example `[%to_yojson: 'a option]` would generate
something like `serialize_option poly_a`.

The fix extends the functions on core types underlying the
%{of,to}_yojson extensions to accept the form `Ptyp_poly`, which was
previously unsupported. On a type of the form `'a . 'a option`, it
would generate the code `fun poly_a -> serialize_option poly_a`. Then
the extensions, called on a core type, start by calling the
Ppx_deriving.strong_type_of_core_type function, which turns `'a
option` into `'a . 'a option`.

Something slightly weird with this approach is that the Ptyp_poly form
cannot be written down by the user: due to limitations of the
extension syntax in current OCaml versions (tested with 4.08),
`[%to_yojson: 'a. 'a option]` is not syntactically correct.