Add missing dependency on result
ppx_deriving_yojson depends on the result compatibility package
but wasn't declaring this dependency explicitly.

This used to work because ppx_deriving also depended on result
but since it will not be the case anymore starting from
ppx_deriving.6.0.0, this needs to be updated.

It's likely that we will drop this dependency for the same reasons
ppx_deriving did but in the meantime it does not cost much to
update our opam file.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>